+++ src/d3.js
@@ -2,7 +2,6 @@ export { select, selectAll } from "d3-selection";
 export { range } from "d3-array";
 export { drag } from "d3-drag";
 export {
-  stack,
   pie,
   arc,
   symbol,
@@ -11,7 +10,7 @@ export {
   symbolTriangle,
 } from "d3-shape";
 export { transition } from "d3-transition";
-export { stratify, cluster, tree, hierarchy } from "d3-hierarchy";
+export { tree, hierarchy } from "d3-hierarchy";
 export { scaleLinear, scaleBand, scaleOrdinal } from "d3-scale";
 
 const colors = (s) => s.match(/.{6}/g).map((x) => `#${x}`);
@@ -20,8 +19,6 @@ export const schemeCategory20b = colors(
 );
 
 export { tsvParseRows } from "d3-dsv";
-export { zoom } from "d3-zoom";
-export { interpolate } from "d3-interpolate";
 export { dispatch } from "d3-dispatch";
 
 export const entries = (object) =>